requirements.txt
Flask==3.0.0
pymysql==1.1.0
python-dotenv==1.0.0
flask-login==0.6.3
bcrypt==4.1.2
